--- account/src/main/kotlin/com/smart/payment/account/AccountApplication.kt
@@ -1,11 +1,0 @@
-package com.smart.payment.account
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class AccountApplication
-
-fun main(args: Array<String>) {
-	runApplication<AccountApplication>(*args)
-}
