+++ api/build.gradle.kts
@@ -1,0 +1,11 @@
+tasks.getByName("bootJar") {
+    enabled = true
+}
+
+tasks.getByName("jar") {
+    enabled = false
+}
+
+dependencies {
+
+}
