+++ gradle.properties
@@ -1,0 +1,6 @@
+## Project configs ##
+projectGroup=com.smart
+
+## Application version ##
+applicationVersion=0.0.1-SNAPSHOT
+
