--- payment/src/main/kotlin/com/smart/payment/PaymentApplication.kt
@@ -1,11 +1,0 @@
-package com.smart.payment
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class PaymentApplication
-
-fun main(args: Array<String>) {
-	runApplication<PaymentApplication>(*args)
-}
