--- payment/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-spring.application.name=payment
