+++ settings.gradle.kts
@@ -1,3 +1,24 @@
 rootProject.name = "fintech"
 
-include("account", "wallet", "payment", "transfer")
+pluginManagement {
+    repositories {
+        mavenCentral()
+        gradlePluginPortal()
+    }
+}
+
+dependencyResolutionManagement {
+    @Suppress("UnstableApiUsage")
+    repositories {
+        mavenCentral()
+        gradlePluginPortal()
+    }
+}
+
+include(
+    "api",
+    "account",
+    "wallet",
+    "payment",
+    "transfer"
+)
