--- transfer/src/main/kotlin/com/smart/payment/transfer/TransferApplication.kt
@@ -1,11 +1,0 @@
-package com.smart.payment.transfer
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class TransferApplication
-
-fun main(args: Array<String>) {
-	runApplication<TransferApplication>(*args)
-}
