+++ wallet/build.gradle.kts
@@ -1,9 +1,3 @@
-plugins {
-    id("org.springframework.boot") version "3.4.1"
-    id("io.spring.dependency-management") version "1.1.7"
-}
-
 dependencies {
-    implementation("org.springframework.boot:spring-boot-starter")
-    testImplementation ("org.springframework.boot:spring-boot-starter-test")
+
 }
